Refactors FHIR extensions for QuestionnaireItem
